rename: response pojo class 패키지 분리
