+++ core/network/src/main/java/com/cheeruplife/core/network/retrofit/seoul/SeoulDataServiceApi.kt
@@ -1,4 +1,5 @@
 package com.cheeruplife.core.network.retrofit.seoul
 
 interface SeoulDataServiceApi {
+    fun getJobCafeList()
 }
