+++ core/network/src/main/java/com/cheeruplife/core/network/retrofit/work24/Work24ServiceApi.kt
@@ -1,8 +1,8 @@
 package com.cheeruplife.core.network.retrofit.work24
 
-import com.cheeruplife.core.network.model.EducationResponse
-import com.cheeruplife.core.network.model.EventResponse
-import com.cheeruplife.core.network.model.RecruitResponse
+import com.cheeruplife.core.network.model.work24.EducationResponse
+import com.cheeruplife.core.network.model.work24.EventResponse
+import com.cheeruplife.core.network.model.work24.RecruitResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
